--- a/package.json
+++ b/package.json
@@ -12,6 +12,7 @@
   "keywords": [
     "linked list",
     "sorted linked list",
+    "skip list",
     "data structure",
     "abstract data type",
     "adt"