Adding support for skip lists in order to improve performance for large datasets.

E:\Test>node index.js 10000
LinkedList insert 10000 strings: 34ms
LinkedList remove 1000 strings: 0ms

E:\Test>node index.js 100000
LinkedList insert 100000 strings: 745ms
LinkedList remove 10000 strings: 167ms
[TestCode.zip](https://github.com/markpete/SortedLinkedList/files/303511/TestCode.zip)